+++ library/src/main/java/me/xiaolei/myroom/library/coverts/base/ToFloatConvert.java
@@ -26,7 +26,7 @@ public abstract class ToFloatConvert<T> extends Convert
     @Override
     public Object cursorToJavaObject(Cursor cursor, int columnIndex)
     {
-        return this.cursorToJavaObject((float) cursor.getFloat(columnIndex));
+        return this.cursorToJavaObject(cursor.getFloat(columnIndex));
     }
 
     /**
